Use Repository.finalize_new_version with cookbook repo_key
Add validation for Repository Versions: A repository version must not have
entries with duplicate repo_keys.

Remove the current check done at publication time.

See upstream: https://pulp.plan.io/issues/3541 and https://pulp.plan.io/issues/5706

Closes #68
