+++ CHANGES/68.feature
@@ -1,0 +1,3 @@
+Add validation to repository versions: A repository version must not have
+entries with duplicate repo_keys. Remove the current check done at publication
+time.
