+++ pulp_cookbook/tests/unit/test_publishing.py
@@ -7,16 +7,16 @@ from unittest.mock import Mock
 from django.test import TestCase
 
 from pulp_cookbook.app.models import CookbookPackageContent
-from pulp_cookbook.app.tasks.publishing import check_repo_version_constraint
+from pulp_cookbook.app.repo_version_utils import check_repo_version_constraint
 
 
-class PublishingCheckRepoVersionConstraintTestCase(TestCase):
+class CheckRepoVersionConstraintTestCase(TestCase):
     """Verify check_repo_version_constraint() method."""
 
-    def new_publication_version_content(self):
-        publication = Mock()
-        publication.repository_version.content = CookbookPackageContent.objects.all()
-        return publication
+    def new_version_content(self):
+        repository_version = Mock()
+        repository_version.content = CookbookPackageContent.objects.all()
+        return repository_version
 
     def test_check_repo_version_constraint_ok(self):
         CookbookPackageContent.objects.create(
@@ -29,7 +29,7 @@ class PublishingCheckRepoVersionConstraintTestCase(TestCase):
             name="c1", version="1.0.1", content_id_type="sha256", content_id="2", dependencies={}
         )
 
-        check_repo_version_constraint(self.new_publication_version_content())
+        check_repo_version_constraint(self.new_version_content())
 
     def test_check_repo_version_constraint_not_ok(self):
         CookbookPackageContent.objects.create(
@@ -48,7 +48,7 @@ class PublishingCheckRepoVersionConstraintTestCase(TestCase):
             name="c1", version="1.0.1", content_id_type="sha256", content_id="1", dependencies={}
         )
         with self.assertRaises(ValueError):
-            check_repo_version_constraint(self.new_publication_version_content())
+            check_repo_version_constraint(self.new_version_content())
 
     def test_check_repo_version_constraint_same_and_different_id_types_not_ok(self):
         CookbookPackageContent.objects.create(
@@ -60,4 +60,4 @@ class PublishingCheckRepoVersionConstraintTestCase(TestCase):
         )
 
         with self.assertRaises(ValueError):
-            check_repo_version_constraint(self.new_publication_version_content())
+            check_repo_version_constraint(self.new_version_content())
